feat: support Java9+ macOS menus
