--- java/src/apps/plaf/macosx/AboutHandler.js
@@ -1,11 +1,0 @@
-/*
- * Set the Java Desktop aboutHandler to handler (passed in script context)
- */
-
-var aboutHandler = new java.awt.desktop.AboutHandler() {
-    handleAbout: function(e) {
-        handler.handleAbout(e);
-    }
-};
-
-java.awt.Desktop.getDesktop().setAboutHandler(aboutHandler);
