--- java/test/jmri/plaf/macosx/Jdk9ApplicationTest.java
@@ -1,18 +1,0 @@
-package jmri.plaf.macosx;
-
-import org.junit.Assert;
-import org.junit.jupiter.api.*;
-
-/**
- * Tests for the Jdk9Application class
- *
- * @author Paul Bender Copyright (C) 2016
- */
-public class Jdk9ApplicationTest  {
-
-   @Test
-   public void testCtor(){
-      Assert.assertNotNull(new Jdk9Application());
-   }
-
-}
